% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopattern.R
\name{geopattern}
\alias{geopattern}
\title{Initiate a geopattern image}
\usage{
geopattern(string, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{string}{the string that will be encoded as an image}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{elementId}{the element-id}
}
\description{
Initiate a geopattern image
}

